BIN files for V1 & simulator V2
