+++ Taewoong/MIPS_operation_simulator.c
@@ -4,28 +4,42 @@
 #include <math.h>
 
 typedef struct {
-	unsigned char inst[4];
-	unsigned char full_bin[33];
-	unsigned char op[7];
-	unsigned char fn[7];
-} BCode;
-
-typedef struct {
 	unsigned char b6_b10[6];
 	unsigned char b11_b15[6];
 	unsigned char b16_b20[6];
 	unsigned char b21_b25[6];
 } BParsed;
 
+typedef struct {
+	unsigned char inst[4];
+	unsigned char full_bin[33];
+	unsigned char op[7];
+	unsigned char fn[7];
+	BParsed bp;
+} BCode;
+
+typedef struct {
+	int num;
+	int data;
+} Register;
+
+typedef struct {
+	long i;
+	long address;
+	BCode bc;
+} Mem_4;
+
+void initReg(Register* reg, Register* PC);
+void initMem(Mem_4* m, long l);
+
 void shell(char* c);
 void readFile(FILE* f, BCode* li);
 
 void inst_config(BCode* list, int len);
 void hex2bin(unsigned char h, char* dd, int r);
-
-void alprint_master(BCode b, int i);
 BParsed parse_inst_bin(BCode b);
 
+void alprint_master(BCode b, int i);
 void alprint_inst_3ar(BCode li, int i, const char* op, BParsed p);
 void alprint_inst_3sh(BCode li, int i, const char* op, BParsed p);
 void alprint_inst_3sl(BCode li, int i, const char* op, BParsed p);
@@ -40,6 +54,33 @@ void alprint_inst_1imm(BCode li, int i, const char* op, BParsed p);
 void alprint_inst_1off(BCode li, int i, const char* op, BParsed p);
 void alprint_unknown(BCode li, int i);
 
+int run_inst_master(Register* r, Register* pc, Mem_4* m, int i);
+
+void run_inst_3ar_sltu(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_slt(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_nor(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_xor(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_or(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_and(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_sub(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_sub(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_add(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3ar_add(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3sh_srav(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3sh_srlv(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3sh_sllv(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3sl_srl(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3sl_sll(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_3sl_sra(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_imm_addi(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_imm_addi(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_imm_slti(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_imm_sltiu(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_imm_andi(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_imm_ori(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_imm_xori(Register* r, Register* pc, Mem_4* m, int i);
+void run_inst_1imm_lui(Register* r, Register* pc, Mem_4* m, int i);
+ 
 long charbin2long(unsigned char* c);
 int charbin2int(unsigned char* c);
 char* mygets(char* s);
@@ -47,12 +88,23 @@ char* mygets(char* s);
 int main() {
 	char command[100];
 	char* object;
+	int i = 0;
+	int list_len = 0;
+
+	Register reg[32];
+	Register PC;
+	initReg(reg, &PC);
+	BCode inst_list[64];
+	Mem_4 memory[64];
 
 	while (1) {
 		shell(command);
 
-		if (strcmp(command, "exit") == 0)
+		if (strcmp(command, "exit") == 0) {
+			free(inst_list);
+			free(memory);
 			break;
+		}
 
 		else if (strlen(command) == 0)
 			continue;
@@ -61,26 +113,25 @@ int main() {
 			object = strtok(command, " ");
 			if (!strcmp(object, "read")) {
 				if (object = strtok(NULL, " ")) {
-					int list_len = 0;
+					list_len = 0;
+					i = 0;
 					unsigned char buffer[4];
 					FILE* len = fopen(object, "rb");
 					if (!len)
 						continue;
-
 					while (fread(buffer, sizeof(unsigned char), 4, len) != 0)
 						list_len++;
 					fclose(len);
 
 					FILE* f = fopen(object, "rb");
-					BCode* inst_list = (BCode*)malloc(sizeof(BCode) * list_len);
-
 					readFile(f, inst_list);
 					inst_config(inst_list, list_len);
 
-					for (int p = 0; p < list_len; p++)
-						alprint_master(inst_list[p], p);
+					for (i; i < list_len; i++) {
+						inst_list[i].bp = parse_inst_bin(inst_list[i]);
+						alprint_master(inst_list[i], i);
+					}
 
-					free(inst_list);
 					fclose(f);
 					continue;
 				}
@@ -89,10 +140,84 @@ int main() {
 					continue;
 				}
 			}
+			else if (strcmp(object, "loadinst") == 0) {
+				if (object = strtok(NULL, " ")) {
+					list_len = 0;
+					unsigned char buffer[4];
+					FILE* len = fopen(object, "rb");
+					if (!len)
+						continue;
+					while (fread(buffer, sizeof(unsigned char), 4, len) != 0)
+						list_len++;
+					fclose(len);
+
+					FILE* f = fopen(object, "rb");
+					initMem(memory, list_len);
+					readFile(f, inst_list);
+					inst_config(inst_list, list_len);
+
+					for (i = 0; i < list_len; i++) {
+						inst_list[i].bp = parse_inst_bin(inst_list[i]);
+						memory[i].bc.bp = parse_inst_bin(inst_list[i]);
+					}
+					for (i = 0; i < list_len; i++) 
+						memory[i].bc = inst_list[i];
+					
+					fclose(f);
+					continue;
+				}
+				else {
+					printf("Please choose a file to read...\n");
+					continue;
+				}
+			}
+			else if (strcmp(object, "run") == 0) {
+				if (object = strtok(NULL, " ")) {
+					int state = 0;
+					i = 0;
+					initReg(reg, &PC);
+					while (1) {
+						if (i + 1 > atoi(object))
+							break;
+						state = run_inst_master(reg, &PC, memory, i++);
+						
+						if (state == -1) {
+							printf("unknown instruction\n");
+							i--;
+							break;
+						}
+					}
+					printf("Executed %d instructions\n", i);
+				}
+			}
+			else if (strcmp(object, "registers") == 0) {
+				for (i = 0; i < 32; i++)
+					printf("$%d: 0x%08x\n", i, reg[i].data);
+				printf("PC: 0x%08x\n", PC.data);
+			}
 			else continue;
 		}
 	}
 	return 0;
+}
+
+void initReg(Register* reg, Register* PC) {
+	int i = 0;
+	PC->num = -1;
+	PC->data = 0;
+	for (i; i < 32; i++) {
+		reg[i].num = i;
+		reg[i].data = 0;
+	}
+}
+
+void initMem(Mem_4* m, long l) {
+	int i = 0;
+	for (i; i < l; i++) {
+		strcpy(m[i].bc.full_bin, "11111111111111111111111111111111");
+		m[i].i = i;
+		m[i].address = m[i].i * 4;
+	}
 }
 
 void shell(char* c) {
@@ -117,7 +242,7 @@ void inst_config(BCode* li, int len) {
 	int i = 0;
 	for (i; i < len; i++) {
 		int j = 0;
-		for(j; j < 33; j++)
+		for (j; j < 33; j++)
 			li[i].full_bin[j] = '\0';
 		for (j = 0; j < 4; j++)
 			hex2bin(li[i].inst[j], li[i].full_bin, 0);
@@ -130,7 +255,6 @@ void inst_config(BCode* li, int len) {
 		li[i].fn[k] = '\0';
 	}
 }
-
 
 void hex2bin(unsigned char h, char* dd, int r) {
 	if (r == 0) {
@@ -195,105 +319,104 @@ void hex2bin(unsigned char h, char* dd, int r) {
 }
 
 void alprint_master(BCode b, int i) {
-	BParsed bp = parse_inst_bin(b);
 	if (!strcmp(b.op, "000000")) {
 		if (!strcmp(b.fn, "101011")) // SLTU
-			alprint_inst_3ar(b, i, "sltu", bp);
+			alprint_inst_3ar(b, i, "sltu", b.bp);
 		else if (!strcmp(b.fn, "101010")) // SLT
-			alprint_inst_3ar(b, i, "slt", bp);
+			alprint_inst_3ar(b, i, "slt", b.bp);
 		else if (!strcmp(b.fn, "100111")) // NOR
-			alprint_inst_3ar(b, i, "nor", bp);
+			alprint_inst_3ar(b, i, "nor", b.bp);
 		else if (!strcmp(b.fn, "100110")) // XOR
-			alprint_inst_3ar(b, i, "xor", bp);
+			alprint_inst_3ar(b, i, "xor", b.bp);
 		else if (!strcmp(b.fn, "100101")) // OR
-			alprint_inst_3ar(b, i, "or", bp);
+			alprint_inst_3ar(b, i, "or", b.bp);
 		else if (!strcmp(b.fn, "100100")) // AND
-			alprint_inst_3ar(b, i, "and", bp);
+			alprint_inst_3ar(b, i, "and", b.bp);
 		else if (!strcmp(b.fn, "100011")) // SUBU
-			alprint_inst_3ar(b, i, "subu", bp);
+			alprint_inst_3ar(b, i, "subu", b.bp);
 		else if (!strcmp(b.fn, "100010")) // SUB
-			alprint_inst_3ar(b, i, "sub", bp);
+			alprint_inst_3ar(b, i, "sub", b.bp);
 		else if (!strcmp(b.fn, "100001")) // ADDU
-			alprint_inst_3ar(b, i, "addu", bp);
+			alprint_inst_3ar(b, i, "addu", b.bp);
 		else if (!strcmp(b.fn, "100000")) // ADD
-			alprint_inst_3ar(b, i, "add", bp);
+			alprint_inst_3ar(b, i, "add", b.bp);
 		else if (!strcmp(b.fn, "000111")) // SRAV
-			alprint_inst_3sh(b, i, "srav", bp);
+			alprint_inst_3sh(b, i, "srav", b.bp);
 		else if (!strcmp(b.fn, "000110")) // SRLV
-			alprint_inst_3sh(b, i, "srlv", bp);
+			alprint_inst_3sh(b, i, "srlv", b.bp);
 		else if (!strcmp(b.fn, "000100")) // SLLV
-			alprint_inst_3sh(b, i, "sllv", bp);
+			alprint_inst_3sh(b, i, "sllv", b.bp);
 		else if (!strcmp(b.fn, "000010")) // SRL
-			alprint_inst_3sl(b, i, "srl", bp);
+			alprint_inst_3sl(b, i, "srl", b.bp);
 		else if (!strcmp(b.fn, "000000")) // SLL
-			alprint_inst_3sl(b, i, "sll", bp);
+			alprint_inst_3sl(b, i, "sll", b.bp);
 		else if (!strcmp(b.fn, "000011")) // SRA
-			alprint_inst_3sl(b, i, "sll", bp);
+			alprint_inst_3sl(b, i, "sll", b.bp);
 		else if (!strcmp(b.fn, "011011")) // DIVU
-			alprint_inst_md(b, i, "divu", bp);
+			alprint_inst_md(b, i, "divu", b.bp);
 		else if (!strcmp(b.fn, "011010")) // DIV
-			alprint_inst_md(b, i, "div", bp);
+			alprint_inst_md(b, i, "div", b.bp);
 		else if (!strcmp(b.fn, "011001")) // MULTU
-			alprint_inst_md(b, i, "multu", bp);
+			alprint_inst_md(b, i, "multu", b.bp);
 		else if (!strcmp(b.fn, "011000")) // MULT
-			alprint_inst_md(b, i, "mult", bp);
+			alprint_inst_md(b, i, "mult", b.bp);
 		else if (!strcmp(b.fn, "001001")) // JALR
-			alprint_inst_1rs(b, i, "jalr", bp);
+			alprint_inst_1rs(b, i, "jalr", b.bp);
 		else if (!strcmp(b.fn, "010011")) // MTLO
-			alprint_inst_1rs(b, i, "mtlo", bp);
+			alprint_inst_1rs(b, i, "mtlo", b.bp);
 		else if (!strcmp(b.fn, "010001")) // MTHI
-			alprint_inst_1rs(b, i, "mthi", bp);
+			alprint_inst_1rs(b, i, "mthi", b.bp);
 		else if (!strcmp(b.fn, "001000")) // JR
-			alprint_inst_1rs(b, i, "jr", bp);
+			alprint_inst_1rs(b, i, "jr", b.bp);
 		else if (!strcmp(b.fn, "010010")) // MFLO
-			alprint_inst_mf(b, i, "mflo", bp);
+			alprint_inst_mf(b, i, "mflo", b.bp);
 		else if (!strcmp(b.fn, "010000")) // MFHI
-			alprint_inst_mf(b, i, "mfhi", bp);
+			alprint_inst_mf(b, i, "mfhi", b.bp);
 		else if (!strcmp(b.fn, "001100")) // SYSCALL
 			alprint_inst_sys(b, i, "syscall");
 		else
 			alprint_unknown(b, i);
 	}
 	else if (!strcmp(b.op, "000010")) // J
-		alprint_inst_j(b, i, "j", bp);
+		alprint_inst_j(b, i, "j", b.bp);
 	else if (!strcmp(b.op, "000011")) // JAL
-		alprint_inst_j(b, i, "jal", bp);
+		alprint_inst_j(b, i, "jal", b.bp);
 	else if (!strcmp(b.op, "000100")) // BEQ
-		alprint_inst_2off(b, i, "beq", bp);
+		alprint_inst_2off(b, i, "beq", b.bp);
 	else if (!strcmp(b.op, "000101")) // BNE
-		alprint_inst_2off(b, i, "bne", bp);
+		alprint_inst_2off(b, i, "bne", b.bp);
 	else if (!strcmp(b.op, "001000")) // ADDI
-		alprint_inst_imm(b, i, "addi", bp);
+		alprint_inst_imm(b, i, "addi", b.bp);
 	else if (!strcmp(b.op, "001001")) // ADDIU
-		alprint_inst_imm(b, i, "addiu", bp);
+		alprint_inst_imm(b, i, "addiu", b.bp);
 	else if (!strcmp(b.op, "001010")) // SLTI
-		alprint_inst_imm(b, i, "slti", bp);
+		alprint_inst_imm(b, i, "slti", b.bp);
 	else if (!strcmp(b.op, "001011")) // SLTIU
-		alprint_inst_imm(b, i, "sltiu", bp);
+		alprint_inst_imm(b, i, "sltiu", b.bp);
 	else if (!strcmp(b.op, "001100")) // ANDI
-		alprint_inst_imm(b, i, "addi", bp);
+		alprint_inst_imm(b, i, "addi", b.bp);
 	else if (!strcmp(b.op, "001101")) // ORI
-		alprint_inst_imm(b, i, "ori", bp);
+		alprint_inst_imm(b, i, "ori", b.bp);
 	else if (!strcmp(b.op, "001110")) // XORI
-		alprint_inst_imm(b, i, "xori", bp);
+		alprint_inst_imm(b, i, "xori", b.bp);
 	else if (!strcmp(b.op, "001111")) // LUI
-		alprint_inst_1imm(b, i, "lui", bp);
+		alprint_inst_1imm(b, i, "lui", b.bp);
 	else if (!strcmp(b.op, "100000")) // LB
-		alprint_inst_1off(b, i, "lb", bp);
+		alprint_inst_1off(b, i, "lb", b.bp);
 	else if (!strcmp(b.op, "100001")) // LH
-		alprint_inst_1off(b, i, "lh", bp);
+		alprint_inst_1off(b, i, "lh", b.bp);
 	else if (!strcmp(b.op, "100011")) // LW
-		alprint_inst_1off(b, i, "lw", bp);
+		alprint_inst_1off(b, i, "lw", b.bp);
 	else if (!strcmp(b.op, "100100")) // LBU
-		alprint_inst_1off(b, i, "lbu", bp);
+		alprint_inst_1off(b, i, "lbu", b.bp);
 	else if (!strcmp(b.op, "100101")) // LHU
-		alprint_inst_1off(b, i, "lhu", bp);
+		alprint_inst_1off(b, i, "lhu", b.bp);
 	else if (!strcmp(b.op, "101000")) // SB
-		alprint_inst_1off(b, i, "sb", bp);
+		alprint_inst_1off(b, i, "sb", b.bp);
 	else if (!strcmp(b.op, "101001")) // SH
-		alprint_inst_1off(b, i, "sh", bp);
+		alprint_inst_1off(b, i, "sh", b.bp);
 	else if (!strcmp(b.op, "101011")) // SW
-		alprint_inst_1off(b, i, "sw", bp);
+		alprint_inst_1off(b, i, "sw", b.bp);
 	else
 		alprint_unknown(b, i);
 }
@@ -320,7 +443,6 @@ void alprint_inst_3ar(BCode li, int i, const char* op, BParsed p) {
 		i, li.inst[0], li.inst[1], li.inst[2], li.inst[3], op,
 		charbin2int(p.b16_b20), charbin2int(p.b6_b10), charbin2int(p.b11_b15));
 }
-
 void alprint_inst_3sh(BCode li, int i, const char* op, BParsed p) {
 	printf("inst %d: %02x%02x%02x%02x %s $%d, $%d, $%d\n",
 		i, li.inst[0], li.inst[1], li.inst[2], li.inst[3], op,
@@ -401,6 +523,160 @@ void alprint_unknown(BCode li, int i) {
 		i, li.inst[0], li.inst[1], li.inst[2], li.inst[3]);
 }
 
+int run_inst_master(Register* r, Register* pc, Mem_4* m, int i) {
+	int res = 0;
+	// printf("[%d] %s\n", i, m[i].bc.op);
+	if (!strcmp(m[i].bc.op, "000000")) {
+		if (!strcmp(m[i].bc.fn, "101011")) // SLTU
+			run_inst_3ar_sltu(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "101010")) // SLT
+			run_inst_3ar_slt(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100111")) // NOR
+			run_inst_3ar_nor(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100110")) // XOR
+			run_inst_3ar_xor(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100101")) // OR
+			run_inst_3ar_or(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100100")) // AND
+			run_inst_3ar_and(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100011")) // SUBU
+			run_inst_3ar_sub(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100010")) // SUB
+			run_inst_3ar_sub(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100001")) // ADDU
+			run_inst_3ar_add(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "100000")) // ADD
+			run_inst_3ar_add(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "000111")) // SRAV
+			run_inst_3sh_srav(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "000110")) // SRLV
+			run_inst_3sh_srlv(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "000100")) // SLLV
+			run_inst_3sh_sllv(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "000010")) // SRL
+			run_inst_3sl_srl(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "000000")) // SLL
+			run_inst_3sl_sll(r, pc, m, i);
+		else if (!strcmp(m[i].bc.fn, "000011")) // SRA
+			run_inst_3sl_sra(r, pc, m, i);
+		else
+			res = -1;
+	}
+	else if (!strcmp(m[i].bc.op, "001000")) // ADDI
+		run_inst_imm_addi(r, pc, m, i);
+	else if (!strcmp(m[i].bc.op, "001001")) // ADDIU
+		run_inst_imm_addi(r, pc, m, i);
+	else if (!strcmp(m[i].bc.op, "001010")) // SLTI
+		run_inst_imm_slti(r, pc, m, i);
+	else if (!strcmp(m[i].bc.op, "001011")) // SLTIU
+		run_inst_imm_sltiu(r, pc, m, i);
+	else if (!strcmp(m[i].bc.op, "001100")) // ANDI
+		run_inst_imm_andi(r, pc, m, i);
+	else if (!strcmp(m[i].bc.op, "001101")) // ORI
+		run_inst_imm_ori(r, pc, m, i);
+	else if (!strcmp(m[i].bc.op, "001110")) // XORI
+		run_inst_imm_xori(r, pc, m, i);
+	else if (!strcmp(m[i].bc.op, "001111")) // LUI
+		run_inst_1imm_lui(r, pc, m, i);
+	else
+		res = -1;
+
+	pc->data += 4;
+	return res;
+}
+
+void run_inst_3ar_sltu(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = (unsigned)(r[charbin2int(m[i].bc.bp.b6_b10)].data) < (unsigned)(r[charbin2int(m[i].bc.bp.b11_b15)].data) ? 1 : 0;
+}
+void run_inst_3ar_slt(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data < r[charbin2int(m[i].bc.bp.b11_b15)].data ? 1 : 0;
+}
+void run_inst_3ar_nor(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = ~(r[charbin2int(m[i].bc.bp.b6_b10)].data | r[charbin2int(m[i].bc.bp.b11_b15)].data);
+}
+void run_inst_3ar_xor(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data ^ r[charbin2int(m[i].bc.bp.b11_b15)].data;
+}
+void run_inst_3ar_or(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data | r[charbin2int(m[i].bc.bp.b11_b15)].data;
+}
+void run_inst_3ar_and(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data & r[charbin2int(m[i].bc.bp.b11_b15)].data;
+}
+void run_inst_3ar_sub(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data - r[charbin2int(m[i].bc.bp.b11_b15)].data;
+}
+void run_inst_3ar_add(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data + r[charbin2int(m[i].bc.bp.b11_b15)].data;
+}
+void run_inst_3sh_srav(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data >> r[charbin2int(m[i].bc.bp.b11_b15)].data;
+}
+void run_inst_3sh_srlv(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = (unsigned int)r[charbin2int(m[i].bc.bp.b11_b15)].data >> r[charbin2int(m[i].bc.bp.b6_b10)].data;
+}
+void run_inst_3sh_sllv(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = (r[charbin2int(m[i].bc.bp.b6_b10)].data) << r[charbin2int(m[i].bc.bp.b11_b15)].data;
+}
+void run_inst_3sl_srl(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = (unsigned int)(r[charbin2int(m[i].bc.bp.b11_b15)].data) >> (unsigned int)(charbin2int(m[i].bc.bp.b21_b25));
+}
+void run_inst_3sl_sll(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = (r[charbin2int(m[i].bc.bp.b6_b10)].data) << charbin2int(m[i].bc.bp.b21_b25);
+}
+void run_inst_3sl_sra(Register* r, Register* pc, Mem_4* m, int i) {
+	r[charbin2int(m[i].bc.bp.b16_b20)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data >> charbin2int(m[i].bc.bp.b21_b25);
+}
+void run_inst_imm_addi(Register* r, Register* pc, Mem_4* m, int i) {
+	unsigned char targ[17];
+	strcpy(targ, m[i].bc.bp.b16_b20);
+	strcat(targ, m[i].bc.bp.b21_b25);
+	strcat(targ, m[i].bc.fn);
+	r[charbin2int(m[i].bc.bp.b11_b15)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data + (int)(charbin2long(targ));
+}
+void run_inst_imm_slti(Register* r, Register* pc, Mem_4* m, int i) {
+	unsigned char targ[17];
+	strcpy(targ, m[i].bc.bp.b16_b20);
+	strcat(targ, m[i].bc.bp.b21_b25);
+	strcat(targ, m[i].bc.fn);
+	r[charbin2int(m[i].bc.bp.b11_b15)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data < (int)(charbin2long(targ)) ? 1 : 0;
+}
+void run_inst_imm_sltiu(Register* r, Register* pc, Mem_4* m, int i) {
+	unsigned char targ[17];
+	strcpy(targ, m[i].bc.bp.b16_b20);
+	strcat(targ, m[i].bc.bp.b21_b25);
+	strcat(targ, m[i].bc.fn);
+	r[charbin2int(m[i].bc.bp.b11_b15)].data = (unsigned)r[charbin2int(m[i].bc.bp.b6_b10)].data < (unsigned int)(int)(charbin2long(targ)) ? 1 : 0;
+}
+void run_inst_imm_andi(Register* r, Register* pc, Mem_4* m, int i) {
+	unsigned char targ[17];
+	strcpy(targ, m[i].bc.bp.b16_b20);
+	strcat(targ, m[i].bc.bp.b21_b25);
+	strcat(targ, m[i].bc.fn);
+	r[charbin2int(m[i].bc.bp.b11_b15)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data & (int)(charbin2long(targ));
+}
+void run_inst_imm_ori(Register* r, Register* pc, Mem_4* m, int i) {
+	unsigned char targ[17];
+	strcpy(targ, m[i].bc.bp.b16_b20);
+	strcat(targ, m[i].bc.bp.b21_b25);
+	strcat(targ, m[i].bc.fn);
+	r[charbin2int(m[i].bc.bp.b11_b15)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data | (int)(charbin2long(targ));
+}
+void run_inst_imm_xori(Register* r, Register* pc, Mem_4* m, int i) {
+	unsigned char targ[17];
+	strcpy(targ, m[i].bc.bp.b16_b20);
+	strcat(targ, m[i].bc.bp.b21_b25);
+	strcat(targ, m[i].bc.fn);
+	r[charbin2int(m[i].bc.bp.b11_b15)].data = r[charbin2int(m[i].bc.bp.b6_b10)].data ^ (int)(charbin2long(targ));
+}
+void run_inst_1imm_lui(Register* r, Register* pc, Mem_4* m, int i) {
+	unsigned char targ[17];
+	strcpy(targ, m[i].bc.bp.b16_b20);
+	strcat(targ, m[i].bc.bp.b21_b25);
+	strcat(targ, m[i].bc.fn);
+	r[charbin2int(m[i].bc.bp.b11_b15)].data = (unsigned int)(int)(charbin2long(targ)) << 16;
+}
+
 long charbin2long(unsigned char* c) {
 	//printf("========%s=======\n", c);
 	long s = -1 * (c[0] - '0');
@@ -429,7 +705,7 @@ long charbin2long(unsigned char* c) {
 }
 
 int charbin2int(unsigned char* c) {
-	int raw = atol(c);
+	int raw = atoi(c);
 	int i = 0;
 	int d = strlen(c);
 	int res = 0;
@@ -444,7 +720,6 @@ int charbin2int(unsigned char* c) {
 	}
 	return res;
 }
-
 
 char* mygets(char* s) {
 	char buffer;
